Add preliminary subtype test #618
This adds an implementation of subtyping which employs the latest
version of WyAutl and is based on a set of WyRL rules.  This extension
consists only of a conversion procedure from old Automaton to new
Automaton, along with a subtype operator.  This subtype operator can
then be compared with the old one to check for consistency.