Add partial implementation of "Back Propagation" #709
The back propagation phase is useful for moving type information from
sinks to sources.  For example, when performing an assignmnet "x = y" we
want to push the type information of x into the AST for y.  This ensures
that code generation for y can generate the best type, rather than
potentially creating the wrong type and then having to coerce it.