WIP: infer starting bound invariant from context
Identifies integer variables that will always change in the body
of the loop by a simple arithmetic operation (+ or - of constants or
the variable).
The simple arithmetic operation is checked to see if it will make
the variable increase or decrease each loop.
From there determine the value of the variable at the entry of the loop.
With this information an invariant can be created, such as i >= 0 from
setting i to zero before the loop and always increasing i during the
loop.

With this inference, at least While_Valid_{1,3}.whiley can
reduce their proof obligation.