add invariant generator for arrays made from generators
Adds the invariant `|local| == |var|` when local is initialised
with `local = [x; |var|]` and assigns to indexes of `local`
inside the loop.

This is targeted at the map pattern seen in Whiley,
such as making a new array to hold the results of a function
on each element from another array, e.g. `var`.