Fix Array Length copy invariant
Now builds up pairs of arrays from the given context then
finds arrays that have been used in the loop body.
If any of the pair has been re-assigned it is invalidated.
If one of the pair is used in the loop body an invariant
of them remaining equal is added.