Add Array Initialising quantifier loop invariant
Checks if a loop is inited with an expression that does not refer
to it self so it can be used in a `for all` quantifier without ghost
arrays. This is to cover the cases of setting all values in an array
to 0 or some function of index.

The limitations of the current implementation is that not all
expression types are currently support & will result in an error.